Silenced more warnings when building with 7.1 SDK on 64 bit

+++ ABFramework-Examples/Classes/Examples/ABInfiniteViewExample.m
@@ -57,7 +57,7 @@
 #pragma mark - ABInfiniteViewDelegate
 -(void) abInfiniteView:(ABInfiniteView *)infiniteView didSelectViewAtIndex:(NSUInteger)index
 {
-    ABLogInteger(index);
+    ABLogInteger((int)index);
 }
 
 @end
